Renamed SerialGrav -> SerialRecorder, refactoring and cutting legacy object code in SerialGrav
